requirements.txt
streamlit>=1.33
pandas>=2.2
matplotlib>=3.8
python-dotenv>=1.0
google-generativeai>=0.7
graphviz>=0.20
